refactor: `learnerd` commands
