+++ crates/learnerd/src/errors.rs
@@ -13,6 +13,8 @@
 
 use thiserror::Error;
 
+pub type Result<T> = core::result::Result<T, LearnerdErrors>;
+
 /// Errors that can occur during CLI operations.
 ///
 /// This enum wraps various error types from dependencies and the underlying
